Dev/requirements.txt
python-dotenv

ollama

flask
flask-cors

sounddevice
numpy
pydub
pynput

pygame

insanely-fast-whisper

nltk
evaluate
bert-score
torch
requests
absl-py
rouge_score